Add initial API Gateway service
This commit adds the initial code for the API Gateway service. It
currently only contains a simple Lambda proxy service for API
gateway, it doesn't support having a separate Lambda listener on
each resource method. That can be added later, but for now I don't
have anyone who wants it.

This code also changes some other pieces of the codebase that had
to be refactored to take this new service into account.