Add producing/consuming checks to service dependency links.
Previously, we didn't check whether a particular service (like EFS)
was consumable from another service (like API Gateway). Not all
AWS services can be consumed by others, so each service needs to
provide hints about what it produces, as well as what it can
consume. This is provided in the form of output types (such as
environment variables, scripts, etc.), that the service deployer
specifies.

The Handel validation then reads this information from the deployers
as it looks at dependency information, and outputs errors if any
dependencies are specified for incompatible services