This is an initial implementation of the Beanstalk service.
Things seem to be working well minus the ebextensions that can be
embedded in the source directory. I'm having trouble getting them
to execute, but I have the same trouble on a vanilla app that
wasn't created by Handel, so I have a feeling I'm just going crazy
and doing things wrong.