Create security groups inside CloudFormation
This change puts security group creation inside a CloudFormation
template. This is to allow people to easily see at a glance all
the resources for their app by looking at CloudFormation. If they
want to delete an environment, all they need to do is delete the
templates for that env and all resources will go away.