Improve ECS instance-level auto-scaling.
This change adds a Lambda that checks every minute to check whether
clusters need to be scaled up or down. It uses the concept of the
'largest schedulable container' from
http://garbe.io/blog/2017/04/12/a-better-solution-to-ecs-autoscaling/
to decide whether the cluster should scale up, scale down, or stay
put.

Now that this is in place, I've also set the cluster minimum healthy
to 50 percent instead of 0 percent, which will now do deployments
without downtimes. This is still hard-coded for now.