Replace node-s3-client with Python AWS CLI.
This introduces an external dependency on the AWS CLI that can't
be auto-installed with Handel. However, the AWS CLI's S3 sync is
the only good S3 sync option available that I can find. I don't
want to write my own, so for now I'll just require that they have
that installed if they want to use the S3 Static Site service.