Use DependsOn to avoid CloudWatch API Rate Limits
Cloudwatch has very low API Rate Limits. Turns out, it's pretty easy to
hit that limit when setting up lots of application autoscaling rules in
Cloudformation. So, by adding a DependsOn to all of our scaling targets,
we can avoid hitting that rate limit by forcing Cloudformation to only
make one call at a time, rather than parallelizing all of them.

In order to make that happen, I've also restructured the autoscaling
template.  It should produce the same output as before, with added
'DependsOn' attributes, but with a lot less duplication in the
Handlebars template itself. Plus, the new structure makes it a LOT
easier to compute the 'DependsOn' values.