Remove legacy injected env var format.
Now, it's just the DEPENDENCY_NAME_VALUE_NAME format.

We've kept the legacy version around for a while; it's about time we
drop it. This is purely unselfish on my part and has no relation to
the fact that I just hit the 4k limit on total env var length on one of
my lambdas. Nope, definitely not.

Speaking of which, we may want to come up with a strategy for dealing
with the fact that many services have a limit on environment variables.