Tag handel account-wide resources
Fixes #360.

Adds an optional field to the account config file, handel_resource_tags,
which contains tags that Handel will apply to any non-application,
account-wide resources that Handel may create.

This change will not be able to modify any resources that already exist.
Those will either have to be deleted and re-created, or one can manually
add the tags to the cloudformation templates that Handel creates to add
the tags.