Integrate Lerna
Everything should work the same way as before, but it's all Lerna under
the hood.  So, running `npm build` in the root directory will still
work, but it will build all packages using lerna.

Now, sure, right now, we only have one package, but that'll change very,
very soon.

Note that it will be necessary to remove node_modules and re-run npm
install for everything to be nice and proper and happy.