Add a nicer message when Route53 zones are missing (#449)
* Add a nicer message when Route53 zones are missing

This also unifies the message across all places where we're searching
for a zone ID.

It may make more sense to create another method that wraps the
getBestMatchingHostedZone function, which throws if there isn't one.  I
don't necessarily want to do that inside of getBestMatchingHostedZone,
because I can see cases where we'd want to use the falsy-returning
version.  Perhaps a method named 'requireBestMatchingHostedZone'?

* Remove a now-unnecessary TODO