Make static site read-only consumable from other services (#451)
The S3 Static Site service is currently not consumable by other services like Beanstalk or Lambda. This change makes it consumable, but only in a read-only manner. This will help prevent frustration with writing changes to the bucket but then having them overwritten during the next deploy, while still allowing other services to read the contents in the static site.