S3 Static Site - Make CORS work when using CloudFront (#453)
This PR makes CloudFront work with the CORS configurations on S3 buckets.

I didn't end up implementing CORS in the S3 Static Site service because S3 puts a default configuration on buckets that seems fine for static sites. In the spirit of YAGNI, I think I'd rather wait until someone has a good use case to need a different CORS configuration for static sites before implementing it, since it adds complexity to the configuration people have to know about for the service.